@@ -3,8 +3,8 @@ package com.course.example.notifywithemail;
 import android.app.NotificationManager;
 import android.app.NotificationChannel;
 import android.content.Context;
-import android.support.v4.app.NotificationCompat;
-import android.support.v7.app.AppCompatActivity;
+import androidx.core.app.NotificationCompat;
+import androidx.appcompat.app.AppCompatActivity;
 import android.app.PendingIntent;
 import android.content.Intent;
 import android.graphics.Color;
